agrowealth_krishisahay/requirements.txt
streamlit
google-generativeai
python-dotenv
sentence-transformers
faiss-cpu
numpy